refactor: break down App.jsx into modular components and hooks
