+++ .gitignore
@@ -1,2 +1,4 @@
+.vscode/
+
 static/
 uploads/
